feature: refactor login screen and add reusable components
