+++ identity_presentation/src/commonMain/kotlin/net/thechance/mena/identity/presentation/screen/components/AuthScreenContainer.kt
@@ -1,0 +1,59 @@
+package net.thechance.mena.identity.presentation.screen.components
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.ColumnScope
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.systemBarsPadding
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.verticalScroll
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.unit.dp
+import mena.identity_presentation.generated.resources.Res
+import mena.identity_presentation.generated.resources.login_background
+import net.thechance.mena.designsystem.presentation.component.image.MenaImage
+import net.thechance.mena.designsystem.presentation.theme.theme.Theme
+import org.jetbrains.compose.resources.painterResource
+
+@Composable
+fun AuthScreenContainer(
+    modifier: Modifier = Modifier,
+    content: @Composable ColumnScope.() -> Unit
+) {
+    Box(
+        modifier
+            .fillMaxSize()
+            .background(Theme.colorScheme.background.surface)
+    ) {
+        MenaImage(
+            painter = painterResource(Res.drawable.login_background),
+            contentDescription = "background",
+            modifier = Modifier.fillMaxSize(),
+            contentScale = ContentScale.FillBounds
+        )
+        Column(
+            modifier = Modifier
+                .fillMaxSize()
+                .systemBarsPadding()
+                .verticalScroll(rememberScrollState())
+                .background(
+                    Brush.linearGradient(
+                        listOf(
+                            Theme.colorScheme.background.surface,
+                            Theme.colorScheme.background.surface,
+                            Theme.colorScheme.background.surface.copy(alpha = 0.6f),
+                        )
+                    )
+                )
+                .padding(vertical = 24.dp, horizontal = 16.dp),
+            content = content
+        )
+
+
+    }
+}
